2025-11-17a v 6.16 beta 5
---------------------------------------------------
2025-11-17a v 6.16 beta 5 - JJDai - jjdelalandre@orange.fr
---------------------------------------------------
intégration de XoopsFormJason pour stocker les attributs de dtyle dans un seul champ(question->image_style qui rempce plusieur champs image_height, image_shadow, ...)
Ajout de la dificulté d'un quiz
Dans l'admin des question ajout d'un bouton(point d'interrogation" pour acceder au quiz dans l'interface user afin de faciliter la mise en page des slides.
Ajout d'index sur les tables question et quiz pour optimiser les temps de réponse.
ajout du champ 'subject' comme sous catégorie simpliée.
divers corrections et mise en forme
Sur pageEnd, soummission automatique quand il y a un timer et que focer=true

Onglet "Questions" :
    - ajout de la fonction "Ajout d'une page d'information" à une question.
    - ajout du champ "quest_reference_id" pour lier la question et la page d'information
    - L'idée est de faire suivre la question par la bonne réponse avec des explication reprise dans "quest_explanation" de la queestion.
    - A faire recalculer le poids en fonction de cette liaison

Plugin "textMixte" :
    - Transfert des intrus dans les options pour la variante "listbox1"
    - Correction des listes d'intrux pour la variante "litbox2"

+++ htdocs/modules/quizmaker/admin/quiz-delete.php
@@ -26,18 +26,36 @@ use XoopsModules\Quizmaker\Constants;
 use XoopsModules\Quizmaker\Utility;
 //use JANUS;
 
-        $clPerms->checkAndRedirect('delete_quiz', $quizCat_id,'$quizCat_id', "quiz.php?op=list&cat_id={$quizCat_id}");
-
+        $clPerms->checkAndRedirect('delete_quiz', $quizCat_id,'$quizCat_id', "quiz.php?op=list&cat_id={$quizCat_id}", QUIZMAKER_ADMIN_PERM);
+        $delCatId = $quizmakerHelper->getConfig('action_on_del_cat');
+        //verifie si la catégorie de suppression defini dans les options du module existe
+        //si elle n'existe pas retour à la liste
+        if ($delCatId > 0 && !$categoriesHandler->exists($delCatId)) {
+            redirect_header("quiz.php?cat_id={$quizCat_id}", 5, _AM_QUIZMAKER_QUIZ_DEL_CAT_NOT_EXISTS);
+        }
+        //-------------------------------------------------------
 		if (isset($_REQUEST['ok']) && 1 == $_REQUEST['ok']) {
-			if (!$GLOBALS['xoopsSecurity']->check()) {
-				redirect_header('quiz.php', 3, implode(', ', $GLOBALS['xoopsSecurity']->getErrors()));
-			}
-			if ($quizHandler->delete($quizObj)) {
-				redirect_header("quiz.php?cat_id={$quizCat_id}", 3, _AM_QUIZMAKER_FORM_DELETE_OK);
-			} else {
-				$GLOBALS['xoopsTpl']->assign('error', $quizObj->getHtmlErrors());
-			}
+            if ($delCatId > 0 && $quizCat_id != $delCatId) {
+//                 $quizObj->setVar('quiz_cat_id', $delCatId);
+//                 ($quizHandler->insert($quizObj));
+                $quizHandler->updatField('quiz_cat_id', $delCatId, $quizId);
+//echo "quizCat_id = {$quizCat_id} - delCatId = {$delCatId}";exit;        
+                redirect_header("quiz.php?cat_id={$quizCat_id}", 3, _AM_QUIZMAKER_QUIZ_ON_CAT_OK);
+            }else{
+    			if (!$GLOBALS['xoopsSecurity']->check()) {
+    				redirect_header('quiz.php', 3, implode(', ', $GLOBALS['xoopsSecurity']->getErrors()));
+    			}
+    			if ($quizHandler->delete($quizObj)) {
+    				redirect_header("quiz.php?cat_id={$quizCat_id}", 3, _AM_QUIZMAKER_FORM_DELETE_OK);
+    			} else {
+    				$GLOBALS['xoopsTpl']->assign('error', $quizObj->getHtmlErrors());
+    			}
+            }                          
 		} else {
-            $msg =  sprintf(_AM_QUIZMAKER_FORM_SURE_DELETE, $quizObj->getVar('quiz_id'), $quizObj->getVar('quiz_name'));
+            if ($delCatId > 0 && $quizCat_id != $delCatId) {
+                $msg =  sprintf(_AM_QUIZMAKER_SURE_TRANSFERT_QUIZ, $quizObj->getVar('quiz_id'), $quizObj->getVar('quiz_name'));
+            }else{
+                $msg =  sprintf(_AM_QUIZMAKER_FORM_SURE_DELETE, $quizObj->getVar('quiz_id'), $quizObj->getVar('quiz_name'));
+            }
 			xoops_confirm(['ok' => 1, 'quiz_id' => $quizId, 'op' => 'delete'], $_SERVER['REQUEST_URI'], $msg);
 		}
